requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==7.0.1
Django==4.1.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
Faker==15.3.4
flake8==6.0.0
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pycodestyle==2.10.0
pydocstyle==6.2.2
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-env==0.8.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
